requirements.txt
Flask==2.3.2
osmnx==1.9.3
shapely==2.0.3
pandas==2.2.2
requests==2.31.0
mega.py==1.0.8
gunicorn